Move graphviz dot utility behind a feature flag
